requirements.txt
# Thermal Engine - Python Dependencies

# Qt GUI framework
PySide6>=6.5.0

# Image processing
Pillow>=10.0.0

# System monitoring
psutil>=5.9.0

# HID device communication (for LCD display)
hidapi>=0.14.0

# USB device communication (for experimental displays)
pyusb>=1.2.1

# Video background support (optional)
opencv-python>=4.8.0
numpy>=1.24.0

# Version comparison for update checks
packaging>=23.0

# Python.NET for .NET interop
pythonnet>=3.0.0